feat: implementar padrão de resposta JSON para erros de API

+++ src/main/java/io/github/cursodsousa/libraryapi/service/AutorService.java
@@ -1,7 +1,10 @@
 package io.github.cursodsousa.libraryapi.service;
 
+import io.github.cursodsousa.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.cursodsousa.libraryapi.model.Autor;
 import io.github.cursodsousa.libraryapi.repository.AutorRepository;
+import io.github.cursodsousa.libraryapi.repository.LivroRepository;
+import io.github.cursodsousa.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,12 +15,17 @@ import java.util.UUID;
 public class AutorService {
 
     private final AutorRepository repository;
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository) {
+    public AutorService(AutorRepository repository,AutorValidator validator,LivroRepository livroRepository) {
         this.repository = repository;
+        this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
     public Autor salvar(Autor autor){
+        validator.validar(autor);
         return repository.save(autor);
     }
 
@@ -25,7 +33,7 @@ public class AutorService {
         if (autor.getId() == null){
             throw new IllegalArgumentException("Para atualizar é necessário que o autor já esteja salvo na base.");
         }
-
+        validator.validar(autor);
         repository.save(autor);
     }
 
@@ -34,6 +42,11 @@ public class AutorService {
     }
 
     public void deletar(Autor autor){
+        if (possuirLivro(autor)){
+            throw new OperacaoNaoPermitidaException(
+                 "Não é permitido excluir um autor que possui livros cadastrados!"
+            );
+        }
         repository.delete(autor);
     }
 
@@ -52,5 +65,9 @@ public class AutorService {
         }
 
         return repository.findAll();
+    }
+
+    public boolean possuirLivro(Autor autor){
+        return livroRepository.existsByAutor(autor);
     }
 }
